+++ CHANGELOG.md
@@ -15,6 +15,12 @@ notes on GitHub when we make a new release.__
 - Fixes bug where dataflow will hang if a source's `next_awake` is set
   far in the future.
 
+- Adds `stateful_flat_map` operator.
+
+- *Breaking change* Removes `builder` argument from `stateful_map`.
+  Instead, the initial state value is always `None` and you can call
+  your previous builder by hand.
+
 ## v0.18.1
 
 - Changes the default batch size for `KafkaSource` from 1 to 1000 to match