+++ CHANGELOG.md
@@ -9,11 +9,16 @@ For help with updating to new Bytewax versions, please see the
 __Add any extra change notes here and we'll put them in the release
 notes on GitHub when we make a new release.__
 
+## v0.18.2
+
+- Fixes a bug that prevented the deletion of old state in recovery stores.
+
 - Better error messages on invalid epoch and backup interval
   parameters.
 
 - Fixes bug where dataflow will hang if a source's `next_awake` is set
   far in the future.
+
 
 ## v0.18.1